intial implementation for traversals finished. implemented thread safe and immutable seed generator for random level distribution as using thread_rng required mutability and wasn't thread safe.
