+++ helixdb/src/helix_engine/graph_core/mod.rs
@@ -1,3 +1,6 @@
 pub mod graph_core;
 pub mod traversal;
 pub mod traversal_steps;
+
+#[cfg(test)]
+mod traversal_tests;
